requirements.txt
streamlit>=1.36.0
streamlit-authenticator>=0.3.3
pyyaml>=6.0.2
